change the data type for the score in leaderboard_item from int to double

+++ src/main/java/ru/alex/professionalslanguageapi/dto/game/IncrementScoreDto.java
@@ -6,6 +6,6 @@ import jakarta.validation.constraints.NotNull;
 public record IncrementScoreDto(
         @Min(0)
         @NotNull
-        Integer score
+        Double score
 ) {
 }
